Move proof to its own bench
