+++ Cargo.toml
@@ -15,15 +15,13 @@ license-file = "mit-license.md"
 
 [features]
 default = []
-bench = [ "criterion", "proptest" ]
 mimc = [ "zkp-u256" ]
 dylib = [ "wasmer/dylib", "wasmer-engine-dylib", "wasmer-compiler-cranelift" ]
 
 [[bench]]
 name = "criterion"
 harness = false
-path = "criterion.rs"
-required-features = [ "bench", "proptest" ]
+path = "benches/proof.rs"
 
 [dependencies]
 ark-bn254 = { version = "0.3.0" }
@@ -34,7 +32,6 @@ ark-groth16 = { git = "https://github.com/arkworks-rs/groth16", rev = "765817f",
 ark-relations = { version = "0.3.0", default-features = false }
 ark-std = { version = "0.3.0", default-features = false, features = ["parallel"] }
 color-eyre = "0.5"
-criterion = { version = "0.3", optional = true, features = [ "async_tokio" ] }
 ff = { package="ff_ce", version="0.11"}
 hex = "0.4.0"
 hex-literal = "0.3"
@@ -42,7 +39,6 @@ num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
 once_cell = "1.8"
 poseidon-rs = "0.0.8"
 primitive-types = "0.11.1"
-proptest = { version = "1.0", optional = true }
 rand = "0.8.4"
 rayon = "1.5.1"
 serde = "1.0"
@@ -58,7 +54,7 @@ ethers-core = { git = "https://github.com/gakonst/ethers-rs", default-features =
 
 [dev-dependencies]
 bincode = "1.3.3"
-proptest = "1.0"
+criterion = "0.3" 
 rand_chacha = "0.3.1"
 serde_json = "1.0.79"
 tempfile = "3.0"
